+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project are documented in this file.
 
 ## [Unreleased]
 
+## [1.2.4] - 2026-09-16
+
 ### Fixed
 
 - Notes with Windows line endings are handled correctly.